Move MobileNavMenu into components folder
