+++ src/pages/ChoosePlayerPage.jsx
@@ -22,7 +22,7 @@ export default function ChoosePlayerPage() {
       <div className={classes['assign-teams-page-header']}>
         <BackHeaderButton path="/create-pool" />
         <h2>Assign Teams</h2>
-        <NextHeaderButton path="/pool-page" disabled={!areTeamsSelected} />
+        <NextHeaderButton path="/pool-home" disabled={!areTeamsSelected} />
       </div>
       <ChoosePlayerList poolPlayers={pool.players} />
     </div>
