+++ src/pages/PoolPicksPage.jsx
@@ -1,7 +1,7 @@
 import classes from './PoolPicksPage.module.css';
 import PageHeader from '../components/PageHeader';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
-import MobileNavMenu from './MobileNavMenu';
+import MobileNavMenu from '../components/MobileNavMenu';
 import usePool from '../utils/usePool';
 import DisplayTeams from '../components/DisplayTeams';
 import useIsDesktop from '../utils/useIsDesktop';
