+++ src/pages/ReassignTeamsPage.jsx
@@ -1,7 +1,7 @@
 import classes from './ReassignTeamsPage.module.css';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import PageHeader from '../components/PageHeader';
-import MobileNavMenu from './MobileNavMenu';
+import MobileNavMenu from '../components/MobileNavMenu';
 import PlayerHomeProfile from '../components/PlayerHomeProfile';
 import TeamsList from '../components/TeamsList';
 import usePool from '../utils/usePool';
